+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project (attempts to) adhere to [Semantic Versioning](http://semver.org
 - Remove most of the options (just `dereplicate` and `compare` remain)
 - Add greedy clustering support! Both `multiround_primary_clustering` and `greedy_secondary_clustering`
 - Add centrality support; also handle centrality with greedy_secondary_clustering
+- Add --run_tertiary_clustering. This feature runs a final dRep job within the original dRep folder,
+  and adjusts Cdb and Wdb accordingly (see `run_tertiary_clustering` in `d_evaluate.py`)
 
 ## [2.6.2] - 2020-05-28
 - Log information about GenomeInformation when loading it